add:やめたタスクページ実装

+++ backend/resources/views/home.blade.php
@@ -90,17 +90,20 @@
                                             <div class="col-sm-8" style="text-align: left">
                                             </div>
                                         @endif
-                                        <div class="col-sm-2" style="text-align: center;">
-                                            <a href="#" class="nav-link" style="display:inline;padding:3px;color:black"><i class="nav-icon far fa-check-circle"></i></a>
-                                        </div>
-                                        <form method="POST" id="delete_form_{{$todo->getId()}}" action="{{ route('delete_todo')}}">
-                                            @csrf
+                                        {{-- やめたタスクページでは表示しない --}}
+                                        @if ($title != "やめたタスク")
                                             <div class="col-sm-2" style="text-align: center;">
-                                                <input type="hidden" name="id" value={{$todo->getId()}}>
-                                                <input type="hidden" name="csrf_token" value="{{ csrf_token() }}">
-                                                <a href="javascript:delete_form_{{$todo->getId()}}.submit()" class="nav-link" style="display:inline;padding:3px;color:black"><i class="nav-icon fas fa-trash-alt"></i></a>
+                                                <a href="#" class="nav-link" style="display:inline;padding:3px;color:black"><i class="nav-icon far fa-check-circle"></i></a>
                                             </div>
-                                        </form>
+                                            <form method="POST" id="delete_form_{{$todo->getId()}}" action="{{ route('delete_todo')}}">
+                                                @csrf
+                                                <div class="col-sm-2" style="text-align: center;">
+                                                    <input type="hidden" name="id" value={{$todo->getId()}}>
+                                                    <input type="hidden" name="csrf_token" value="{{ csrf_token() }}">
+                                                    <a href="javascript:delete_form_{{$todo->getId()}}.submit()" class="nav-link" style="display:inline;padding:3px;color:black"><i class="nav-icon fas fa-trash-alt"></i></a>
+                                                </div>
+                                            </form>
+                                        @endif
                                     </div>
                                 </div>
                                 <!-- /.card-footer -->
